project 6/AeroGuard_Project/requirements.txt
streamlit>=1.30.0
tensorflow-cpu>=2.15.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
 